Allow many more types in the database
I previously used u8 as a counter for number all the different types that
can be stored in the database. That is now usize. Not that I expect to
hit that limit any time soon. But it was just a wart that was nice to
get rid of.
Also, the refactorings that I did to make this change convenient
were good anyway, so the code looks better now.

+++ type-vault-trait/src/lib.rs
@@ -3,7 +3,7 @@ use std::{any::TypeId, collections::HashMap, hash::*};
 
 pub type ValueId = [u8; 8];
 
-pub type TypeMap = HashMap<TypeId,u8>;
+pub type TypeMap = HashMap<TypeId,usize>;
 
 pub fn value_id_of(data: impl Hash) -> ValueId {
     let mut s = std::collections::hash_map::DefaultHasher::new();
